🔧 Fix workflow again by rewriting Macs building
Some of Golang workflows said no when I tried to build it with `macos-latest` because it can be only builded on `ubuntu-latest` so I decided to write building Presence binaries for Macs from scratch. I hope that will be my latest fix commit related to workflow. Now it should works I hope.